refactor: group zarr utils
