+++ linc_convert/modalities/df/single_slice.py
@@ -21,7 +21,7 @@ from linc_convert.modalities.df.cli import df
 from linc_convert.utils.j2k import WrappedJ2K, get_pixelsize
 from linc_convert.utils.math import ceildiv
 from linc_convert.utils.orientation import center_affine, orientation_to_affine
-from linc_convert.utils.zarr import make_compressor
+from linc_convert.utils.zarr.compressor import make_compressor
 
 ss = App(name="singleslice", help_format="markdown")
 df.command(ss)
