+++ linc_convert/modalities/psoct/single_volume.py
@@ -31,7 +31,7 @@ from linc_convert.modalities.psoct.cli import psoct
 from linc_convert.utils.math import ceildiv
 from linc_convert.utils.orientation import center_affine, orientation_to_affine
 from linc_convert.utils.unit import to_nifti_unit, to_ome_unit
-from linc_convert.utils.zarr import make_compressor
+from linc_convert.utils.zarr.compressor import make_compressor
 
 single_volume = cyclopts.App(name="single_volume", help_format="markdown")
 psoct.command(single_volume)
